Added FileNet Processing to script

+++ shell-script.sh
@@ -21,11 +21,13 @@ connect_to_cyberark () {
 }
 
 if [ $1 = QQ ]; then
-	FILEDST="/opt/FileNet/shared/host/"
+	POSTURL="https://ddd-cpe-qq.validazione.usinet.it/DDMEGABatch/"
+	FILEDST="/home/tudddf3/"
 elif [ $1 = QE ]; then
-	FILEDST="/opt/FileNet/shared/host/"
+	POSTURL="https://ddd-cpe-qe.collaudo.usinet.it/DDMEGABatch/"
+	FILEDST="/home/tudddf3/"
 elif [ $1 = HV ]; then
-	FILEDST="/opt/FileNet/shared/host/"
+	FILEDST="/home/tudddf3/"
 fi
 LOGFILE="sftp-download"
 
@@ -140,6 +142,26 @@ if [ $RC -ne 0 ]; then
 	echo "RETC = ${RC}"
 	echo "Error: file already present on NAS storage or no access to NAS mount"
 	exit $RC
+fi
+
+readarray -d . -t strarr <<<"${FILESRC}"
+
+FILETYPE=${strarr[4]}
+
+#start FileNet processing
+HTTPS_POST_RC=4
+curl -u <user>:<password> -X POST -F "file=${FILEDST}${FILESRC}" -F "type=${FILETYPE}" ${POSTURL}startJob
+
+while [ ${HTTPS_POST_RC} = 4 ]
+do
+	HTTPS_POST_RC=curl -u <user>:<password> -X POST -F "file=${FILEDST}${FILESRC}" -F "type=${FILETYPE}" -w "%{http_code}" ${POSTURL}checkJob
+	sleep 5m
+done
+
+if [ $HTTPS_POST_RC -ne 0 ]; then
+	echo "RETC = ${HTTPS_POST_RC}"
+	echo "Error: the FileNet processing failed"
+	exit $HTTPS_POST_RC
 fi
 
 find ${FILEDST}${LOGFILE}*.log -mtime +10 -type f -delete
